Move testing config files into shared/tests
