--- packages/api/.env.test
@@ -1,8 +1,0 @@
-DATABASE_URL="postgresql://admin:admin@localhost:5435/db-test?schema=public"
-
-POSTGRES_HOST=localhost
-POSTGRES_PORT=5435
-POSTGRES_USER=admin
-POSTGRES_PASSWORD=admin
-POSTGRES_DB=db-test
-PGDATA=/var/lib/postgresql/data/pgdata
